SmartOps AI Flow/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
celery
redis
openai
pydantic